[BE] adapting images to attachments major changes

+++ ndannamo_project/backend/src/main/java/com/example/backend/service/AttachmentService.java
@@ -2,7 +2,11 @@ package com.example.backend.service;
 
 
 import com.example.backend.dto.AttachmentDTO;
+import com.example.backend.exception.ResourceNotFoundException;
+import com.example.backend.exception.UserNotParticipantException;
+import com.example.backend.exception.UserNotParticipantException.UserNotParticipantExceptionCodes;
 import com.example.backend.model.Attachment;
+import com.example.backend.model.Trip;
 import com.example.backend.model.User;
 import com.example.backend.repositories.AttachmentRepository;
 import com.example.backend.repositories.TripRepository;
@@ -19,17 +23,22 @@ public class AttachmentService {
 
     private final AttachmentRepository attachmentRepository;
     private final TripRepository tripRepository;
+    private final UserService userService;
     private final AttachmentMapperImpl attachmentMapper;
 
-    public AttachmentService(AttachmentRepository attachmentRepository, AttachmentMapperImpl attachmentMapper, TripRepository tripRepository) {
+    public AttachmentService(AttachmentRepository attachmentRepository, AttachmentMapperImpl attachmentMapper, TripRepository tripRepository, UserService userService) {
         this.attachmentRepository = attachmentRepository;
         this.attachmentMapper = attachmentMapper;
         this.tripRepository = tripRepository;
+        this.userService = userService;
     }
 
     public List<Attachment> createAttachments(MultipartFile[] files, String email, Trip trip) throws IOException {
         List<Attachment> attachments = new ArrayList<>();
         User logged_user = userService.getUserByEmail(email);
+        if (!TripService.userIsAParticipant(logged_user, trip)) {
+            throw new UserNotParticipantException(logged_user, UserNotParticipantExceptionCodes.CANT_UPLOAD);
+        }
         for (MultipartFile file : files) {
             Attachment attachment = Attachment.builder()
                     .fileName(file.getOriginalFilename())
@@ -42,7 +51,8 @@ public class AttachmentService {
 
             attachments.add(attachment);
         }
-
+        trip.addAttachments(attachments);
+        tripRepository.save(trip);
         return attachmentRepository.saveAll(attachments);
     }
 
@@ -52,12 +62,24 @@ public class AttachmentService {
         return attachmentRepository.findById(id).orElseThrow();
     }
 
-    public AttachmentDTO getAttachmentDataDTObyId(Long id) {
+    public AttachmentDTO getAttachmentDataDTObyId(Long id, String email) {
         Attachment attachment = getAttachment(id);
+        User user = userService.getUserByEmail(email);
+        if (!TripService.userIsAParticipant(user, attachment.getTrip())) {
+            throw new UserNotParticipantException(user, UserNotParticipantExceptionCodes.CANT_RETRIEVE);
+        }
         return attachmentMapper.toDTO(attachment);
     }
 
-    public void deleteAttachment(Long id) {
+    public void deleteAttachment(Long id, String email) {
+        Attachment attachment = getAttachment(id);
+        User user = userService.getUserByEmail(email);
+        Trip trip = attachment.getTrip();
+        if (!TripService.userIsAParticipant(user, attachment)) {
+            throw new UserNotParticipantException(user, UserNotParticipantExceptionCodes.CANT_DELETE); //FIXME: gestire permessi per eliminazione
+        }
+        attachment.removeAttachment(attachment);
+        tripRepository.save(attachment);
         attachmentRepository.deleteById(id);
     }
